requirements.txt
appdirs==1.4.3
astroid==2.3.2
attrs==19.3.0
black==19.3b0
Click==7.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pendulum==2.0.5
pkg-resources==0.0.0
pylint==2.4.3
python-dateutil==2.8.0
pytzdata==2019.3
PyYAML==5.1.2
six==1.12.0
toml==0.10.0
transmissionrpc==0.11
typed-ast==1.4.0
wrapt==1.11.2
